requirements.txt
langchain
python-dotenv
InstructorEmbedding
sentence_transformers==2.2.2
tiktoken
chromadb
flask
flask-cors
ibm-watsonx-ai>=0.2.2
langchain-ibm
